mxcfb: add Plastic Logic E-Paper power control
* add CONFIG_FB_MXC_EPDC_PL_HARDWARE
* add basic all-in-one power control (CPLD, DAC, HVPMIC)
* support enable and disable to turn the power on/off

Notes: This is an interim solution.

This uses the Maxim MAX17135 high-voltage e-paper PMIC internally,
but does not yet use the existing regulator driver.  This needs to
be improved and also register itself as a regulator driver as
well as using the MAX17135 driver.

The VCOM also needs to be configured at run-time (sysfs?).

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>