mxc_epdc_fb: accept an array of VCOM voltages
The vcom modparam can now accept a comma-separated
list of VCOM voltages.  They are also passed on to
the pl_hardware instance (only the first is used
for now) via mxc_epdc_plhw_config.

The number of VCOM voltages is assumed to be equal to
the number of power supplies as managed by pl_hardware.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>