mxc_epdc_pl_hardware: add PSU I2C switch
* split the HVPMIC and DAC instances into a new
  pl_hardware_psu structure
* add 2 instances of pl_hardware_psu into mxc_epdc_pl_hardware
* add functions to access devices on either I2C branch
* fix PSU i2c_sel_id value (was not initialised)
* add power up/down steps for secondary power supply
* remove unnecessary printk messages
* fix secondary PSU CPLD flags for HVEN1 and COM_SW_CLOSE1
* invert the logic for the I2C switch on the CPLD fast
  interface (looks like there was an error in the doc or
  something changed since older CPLD firmware)

Note: It's possible in principle to turn on the 2 PSUs in
parallel to reduce the latency, although this may cause some
other issues in hardware which haven't been investigated yet.  So
the current solution is to turn them on and off one after the
other.

Note: There is a small chance that some other driver is going
through an I2C transation while the switch is occuring, which
will cause I2C errors due to the devices receiving partial I2C
packets.  To avoid this issue, this driver should lock the I2C
bus driver while making the switch.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>