mx_epdc_pl_hardware: fix ADC reading + tweak last_v_in
The ADC input contains a low-pass filter which helps stabilize
the measurement but also restricts the speed at which ADC
measurements can be made.  To avoid reading invalid measurements,
a 2ms delay is added before reading from the ADC.

Also the last voltage fed into the DAC is now kept as it was
before being changed in the last calibration loop iteration.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>