BeagleBone: add SPI work-around for PL hardware boards
Some Plastic Logic hardware boards such as the Hummingbird
Z1.3 and Robin do not use the SPI MISO/MOSI in the
conventional way.  This change adds a new config option
CONFIG_MODELF_SWAP_SPI_IO to invert them, which will break
standard configurations.  This is an interim fix and should
be reverted when all these boards are deprecated.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>