i2c-sc18is60x: keep shared SPI hack self-contained
Instead of relying on an other driver to define the shared
symbols, move everything back to the sc18is60x driver.  This
boils down to a statically allocated mutex and a pointer to the
shared SPI device.

The SPI configuration is also altered and restored while being
used by the sc18is60x driver (needs MODE_3).

+ reduce delay to 100us in write-then-read sequence
+ fix log message typo

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>