mx50_rdp: Modify pl videomode settings to support type 11 display.
Type 11 displays need to drive all 480 lines on each source chip, so we
need to set an x resolution of 1440 pixels. The first and last 80 lines
of this 1440 will not be visible, so we need to introduce an offset
to avoid losing the top 80 lines of the framebuffer when updating the
display.

Signed-off-by: Andrew Cox <andrew.cox@plasticlogic.com>