mxc_epdc_fb: add EPDC left and right border configuration
On some display designs, the source driver outputs are not all
connected to physical lines.  This can now be dealt with as long
as all the source lines are contiguous.  The number of floating
source outputs on the left and right sides can be configured with
the mxc_epdc_fb_mode.left_boarder and right_border values.

This is transparent to applications as the framebuffer resolution
doesn't change and is still as defined by the video
mode (reflecting the active area resolution).  The PXP
destination address is offset by the value of the left_border,
and update regions are updated accordingly.  So in practice, when
the left_border is non-zero, when updating the full active area
the EPDC logically generates a regional update and ignores the
unused source lines (driven with nulls).

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>