mx50_rdp: update left_border and right_border for MXEI1480
The MXEI1480 source driver chips is now the standard part on
Plastic Logic 1280x960 displays.  It has a fixed number of 480
outputs so the first and last 80 are not connected to actual
display lines.  This is reflected in the left_border and
right_border values.  Other display configurations have these
values set to 0.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>