mxc_epdc_pl_hardware: reinitialise HVPMIC after initial update.
Currently when using Type11 displays there is a big voltage drop on
the first display update, which can trigger a reset of the HVPMIC.
This is especially likely when using dual/tiled displays. As a temp-
orary workaround to this, we will now set up the HVPMIC for a second
time after the first update.

Signed-off-by: Andrew Cox <andrew.cox@plasticlogic.com>