mxc_epdc_pl_hardware: modify HVPMIC reset work-around
Always reload the HVPMIC timings during power-down when using
power sequence SEQ1.  This is to work around the issue when the
HVPMIC resets during a display update due to the VMAIN voltage
falling under its internal 2.6V threshold.  This is a temporary
solution until the root cause of the problem is fixed.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>