S1D135x1: transfer more than one line of image data at a time
Instead of working on a single line at a time, increase the image
pool size to be up to 32KB.  This can contain about 40 lines of
400 16-bit pixels instead of one.  The number of lines is
determined to use the maximum number of bytes in the pool.

Then modify the code that prepares and sends the image data to
use all the lines in one block, which greatly reduces the SPI
transfer overhead.  The overall time to send the image data is
typically divided by 2.

+ print message with number of lines and size of image pool

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>