mxc_epdc_pl_hardware: apply new Maxim HVPMIC timings
Replace Maxim HVPMIC timings for power on and off and for both
SEQ_0 (Type4, Type10) and SEQ_1 (Type11).  This is to ensure that
no more than one power rail is being turned on at any given time,
to minimise the peak inrush current otherwise caused by
overlapping ramps.  The power off timings are also modified to
turn the rails in the reverse order than during power on.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>