pl_hardware: add support for TI TPS65185 HVPMIC + HB Z6-Z7
Add code to initialise and use the TI TPS65185 HVPMIC and its
internal DAC to generate the VCOM voltage.  This can be used on
PL Hummingbird Z6.x and Z7.x which have this TI HVPMIC.

Add a hvpmic_id field to the pl_hardware_config structure to let
the caller choose which HVPMIC to use.  The discrete VCOM DAC is
used only with the Maxim MAX17135 HVPMIC.

The VCOM calibration is not enabled when using HB Z6-Z7.

Also add the GPIOs for the Z6-Z7 boards to the list, and review
the GPIO handling functions to make the code more maintainable
with the Z1.3 and Z5.x GPIOs using a pl_hardware_gpio_init_data
table.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>