S1D135x1: implement SPI-I2C bridge driver interface
Populate the s1d135x1_spi and move the pl_hardware initialisation
only after the SPI-I2C bridge driver is initialised as it may
need it if enabled.

Use i2c-5 when the SPI-I2C bridge driver is enabled, or i2c-3
otherwise.  This still needs to be moved to the platform data.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>