S1D135x1: add optional GPIOs for reset and power enable
Add basic support for the `reset' and `pwr_en' GPIOs as
configured via the platform data.

Some platforms have dedicated pins to reset the controller and to
enable the power of the display electronics.  They need to be
driven high in order to power it and not hold the controller in
reset.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>