S1D135x1: add left and right border parameters
Add parinfo->left_border and ->right_border which represent the
number of source lines on either side of the display that are not
part of the active area.  The framebuffer is then created to
match the active area, and the pixels outside this are ignored by
the software.

This is currently not fully configurable from user-side but is
hard-coded for PL panel types.  For Type16, the left border is
set to 80.  For Type17, the right border is set to 7.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>