S1D135x1: add SPI clock frequency option via sysfs
Add `spi_freq_hz' option to the `setop' command on the `control'
sysfs interface.  This is then used to override the default SPI
frequency set on the platform, in HZ.  All SPI transactions in
the driver then use this frequency.  Setting it to 0 reverts to
the platform default value.

This is stored in modelffb_par.opt.spi_freq_hz.

+ simplify syntax when sending only one SPI packet

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>