modelffb: convert panel_type param into display_type option
Some display hardware contains an EEPROM with the display type,
which can only be accessed when the power is turned on.  The
display driver turns the power on, which makes it impossible to
read the display type from the EEPROM before loading the driver,
so this doesn't work as a module parameter.

Instead, the panel_type parameter can now be set via the control
set_opt command, and is also renamed to display_type for
consistency.  Like interleaved_sources, it needs to be set before
the chip is initialised as it has an impact on the frame buffer
size and display resolution registers.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>