pl_hardware: add static init/free functions
Add pl_hardware_static_{init|free} to allocate and free static
resources.  This basically configures the GPIOs, which as a side
effect can turn the power on when needed by the modelffb driver.

This is still an interim solution until the pl_hardware module
becomes a separate driver with some dedicated platform data.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>