pl_hardware: add init_vcom_mv configuration
Add pl_hardware_config.init_vcom_mv to contain the initial VCOM
value.  If set to PLHW_INVALID_VCOM, then the default (average)
initial VCOM voltage value will be used as before.  This allows
the VCOM voltage to be guaranteed to be correct the first time
the display power is being powered on.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>