modelffb: fix early I2C support when using SPI bridge
When the SPI-I2C bridge of the S1D135x1 is being used, enable it
early (befoew the init code and waveform are loaded) so I2C
devices which may contain the waveform data such as an EEPROM can
be accessed.

Also add a separate mutex for the SPI-I2C bridge to ensure a
consistent state around the init_done value and other members of
the s1d135x1_spi API structure.

+ tidy-up buggy code in _store_init_code
+ use pdev->dev when parinfo is not necessarily initialised
+ tweak log messages

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>