I2C_S1D135X1: remove dependency on FB_MODELF to be built-in
The i2c-s1d135x1 driver will wait until another driver provides
it with an SPI device to use, and this is typically but not
necessarily the modelffb driver.  Also, removing the dependency
means the modelffb can be built as a loadable module while the
i2c-s1d135x1 driver can be built-in.

Signed-off-by: Guillaume Tucker <guillaume.tucker@plasticlogic.com>